+++ pkg/git/provider_github.go
@@ -8,10 +8,12 @@ import (
 
 	g "github.com/argoproj-labs/argocd-autopilot/pkg/git/github"
 
-	gh "github.com/google/go-github/v35/github"
+	gh "github.com/google/go-github/v43/github"
 )
 
-//go:generate mockery --dir github --all --output github/mocks --case snake
+//go:generate mockgen -destination=./github/mocks/repos.go -package=mocks -source=./github/repos.go Repositories
+//go:generate mockgen -destination=./github/mocks/users.go -package=mocks -source=./github/users.go Users
+
 type github struct {
 	opts         *ProviderOptions
 	Repositories g.Repositories
@@ -20,7 +22,7 @@ type github struct {
 
 func newGithub(opts *ProviderOptions) (Provider, error) {
 	var (
-		c *gh.Client
+		c   *gh.Client
 		err error
 	)
 
@@ -50,13 +52,14 @@ func newGithub(opts *ProviderOptions) (Provider, error) {
 	return g, nil
 }
 
-func (g *github) CreateRepository(ctx context.Context, opts *CreateRepoOptions) (string, error) {
-	authUser, res, err := g.Users.Get(ctx, "") // get authenticated user details
+func (g *github) CreateRepository(ctx context.Context, orgRepo string) (string, error) {
+	opts, err := getDefaultRepoOptions(orgRepo)
 	if err != nil {
-		if res.StatusCode == 401 {
-			return "", ErrAuthenticationFailed(err)
-		}
+		return "", nil
+	}
 
+	authUser, err := g.getAuthenticatedUser(ctx)
+	if err != nil {
 		return "", err
 	}
 
@@ -82,4 +85,72 @@ func (g *github) CreateRepository(ctx context.Context, opts *CreateRepoOptions) 
 	}
 
 	return *r.CloneURL, err
+}
+
+func (g *github) GetAuthor(ctx context.Context) (username, email string, err error) {
+	authUser, err := g.getAuthenticatedUser(ctx)
+	if err != nil {
+		return
+	}
+
+	username = authUser.GetName()
+	if username == "" {
+		username = authUser.GetLogin()
+	}
+
+	email = authUser.GetEmail()
+	if email == "" {
+		email = g.getEmail(ctx)
+	}
+
+	if email == "" {
+		email = authUser.GetLogin()
+	}
+
+	return
+}
+
+func (g *github) getAuthenticatedUser(ctx context.Context) (*gh.User, error) {
+	authUser, res, err := g.Users.Get(ctx, "")
+	if err != nil {
+		if res.StatusCode == 401 {
+			return nil, ErrAuthenticationFailed(err)
+		}
+
+		return nil, err
+	}
+
+	return authUser, nil
+}
+
+func (g *github) getEmail(ctx context.Context) string {
+	emails, _, err := g.Users.ListEmails(ctx, &gh.ListOptions{
+		Page:    0,
+		PerPage: 10,
+	})
+	if err != nil {
+		return ""
+	}
+
+	var email *gh.UserEmail
+	for _, e := range emails {
+		if e.GetVisibility() != "public" {
+			continue
+		}
+
+		if e.GetPrimary() && e.GetVerified() {
+			email = e
+			break
+		}
+
+		if e.GetPrimary() {
+			email = e
+		}
+
+		if e.GetVerified() && !email.GetPrimary() {
+			email = e
+		}
+	}
+
+	return email.GetEmail()
 }